+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+### unreleased
+- Remove unused Minterms method of Kmap
+
 ### [v0.1.1](https://github.com/noah-friedman/kmap/tree/v0.1.1)
 - [Error messages](https://github.com/noah-friedman/kmap/blob/v0.1.1/kmap.go) explain what went wrong better